+++ libs/nestjs-gridfs-mongodb/src/index.ts
@@ -1,4 +1,11 @@
 export { GridfsModule } from './gridfs.module';
 export { GridfsService } from './gridfs.service';
-export { GridfsFile, GridfsFileMetadata, GridfsFileBuffer, GridfsGetFileOptions } from './gridfs.types.interface';
+export { 
+    GridfsFile, 
+    GridfsFileMetadata,
+    GridfsFileBuffer, 
+    GridfsGetFileOptions, 
+    GridfsUploadFileResponse, 
+    GridfsDeleteFileResponse 
+} from './gridfs.types.interface';
 export { GridfsConfig, GridfsConfigMetadataIndex } from './gridfs.config.interface';