+++ src/app.controller.ts
@@ -1,6 +1,6 @@
 import { Body, Controller, Param, Post, UploadedFile, UseInterceptors } from "@nestjs/common";
 import { FileInterceptor } from "@nestjs/platform-express";
-import { GridfsFile, GridfsFileMetadata, GridfsGetFileOptions, GridfsService } from "@app/nestjs-gridfs-mongodb";
+import { GridfsDeleteFileResponse, GridfsFile, GridfsFileMetadata, GridfsGetFileOptions, GridfsService, GridfsUploadFileResponse } from "@app/nestjs-gridfs-mongodb";
 
 @Controller('nestjs-gridfs-mongodb')
 export class AppController {
@@ -13,13 +13,12 @@ export class AppController {
     @UploadedFile() file: Express.Multer.File,
     @Param('bucketName') bucketName: string,
     @Body() data: any,
-  ): Promise<boolean> {
-    // You can expect a single file (File) or an array of files (File[])
+  ): Promise<GridfsUploadFileResponse[]> {
     // The body with metadata is optional, and it will be required to pass in text format (JSON.stringify // JSON.parse)
     // If you don't provide metadata, the file will be uploaded with default metadata with mimetype property value
     return await this.gridfsService.uploadFiles(
       bucketName, 
-      file, 
+      [file], 
       data.body 
         ? JSON.parse(data.body.toString()) as GridfsFileMetadata
         : undefined
@@ -30,23 +29,21 @@ export class AppController {
   async getFiles(
     @Param('bucketName') bucketName: string,
     @Body() options: GridfsGetFileOptions,
-  ): Promise<GridfsFile> {
+  ): Promise<GridfsFile[]> {
     /* 
       You provide options to manage query filter.
       Options is an object with the following properties:
       - filter: Object (Object with the properties to filter the documents, same work as moongose find filter)
       - includeBuffer: boolean (Will return the GridfsFile object with the buffer property, buffer includes the data and base64 of the file)
-      - single: boolean (Will return a single document (GridfsFile) or an array of documents (GridfsFile[]))
     */
-    return await this.gridfsService.getFiles(bucketName, options) as GridfsFile;
+    return await this.gridfsService.getFiles(bucketName, options);
   }
 
   @Post('deleteFiles/:bucketName')
   async deleteFiles(
     @Param('bucketName') bucketName: string,
     @Body() data: any,
-  ): Promise<boolean> {
-    // You can expect a single id (string) or an array of ids (string[])
+  ): Promise<GridfsDeleteFileResponse> {
     return await this.gridfsService.deleteFiles(bucketName, data._ids ?? []);
   }
 }