+++ src/app.module.ts
@@ -1,12 +1,9 @@
 import { Module } from '@nestjs/common';
-import { MongooseModule } from '@nestjs/mongoose';
 import { AppController } from './app.controller';
 import { GridfsModule } from '@app/nestjs-gridfs-mongodb';
  
 @Module({
   imports: [
-    MongooseModule.forRoot('mongodb://localhost:27017/nestjs-gridfs-mongodb'), 
-    
     // Simple register with config object
     GridfsModule.register({
       bucketNames: ['client-files', 'worker-files'],