+++ CHANGELOG.md
@@ -9,6 +9,15 @@
 ## [10.0.3] - 2025-04-10
 
 ### Added
+- Add GridfsUploadFileResponse type and modify update response.
+- Add GridfsDeleteFileResponse type and modify delete response.
+
+### Deprecated
+- Update GridfsGetFileOptions type with deleted single property.
+
+## [10.0.3] - 2025-04-10
+
+### Added
 - Add support for multiple buckets.
 - Refactor connection and add support for multiple connections.